Security advisory after snyk test
After conducting a security scan using Snyk, potential security vulnerabilities were discovered in this application. The first vulnerability is Regular Expression Denial of Service (ReDoS), which is a type of security issue that can exploit the way regular expressions are processed and lead to a denial-of-service (DoS) attack. The second vulnerability is Server-side Request Forgery (SSRF), which can enable an attacker to send requests from a vulnerable server to other systems, both internal and external.