After running a security scan using Snyk, it has been discovered that this application may be vulnerable to certain security risks. One such vulnerability is Regular Expression Denial of Service (ReDoS) is a type of security vulnerability that can cause a denial-of-service (DoS) attack by exploiting the way that regular expressions are processed. Another vulnerability is Server-side Request Forgery (SSRF) is a type of security vulnerability that can allow an attacker to make requests from a vulnerable server to other internal or external systems.
After running a security scan using Snyk, it has been discovered that this application may be vulnerable to certain security risks. One such vulnerability is Regular Expression Denial of Service (ReDoS) is a type of security vulnerability that can cause a denial-of-service (DoS) attack by exploiting the way that regular expressions are processed. Another vulnerability is Server-side Request Forgery (SSRF) is a type of security vulnerability that can allow an attacker to make requests from a vulnerable server to other internal or external systems.